pyproject.toml
[project]
name = "gh-transcript-fetch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "openai>=1.61.1",
    "pydub>=0.25.1",
    "pygithub>=2.5.0",
    "python-fasthtml>=0.12.1",
    "pytube>=15.0.0",
    "pytz>=2025.1",
    "youtube-transcript-api>=0.6.3",
]
